+++ arch/arm64/core/CMakeLists.txt
@@ -17,6 +17,7 @@ zephyr_library_sources(
 
 zephyr_library_sources_ifdef(CONFIG_FPU_SHARING fpu.c fpu.S)
 zephyr_library_sources_ifdef(CONFIG_ARM_MMU mmu.c mmu.S)
+zephyr_library_sources_ifdef(CONFIG_ARM_MPU cortex_r/arm_mpu.c)
 zephyr_library_sources_ifdef(CONFIG_USERSPACE userspace.S)
 zephyr_library_sources_ifdef(CONFIG_GEN_SW_ISR_TABLE isr_wrapper.S)
 zephyr_library_sources_ifdef(CONFIG_IRQ_OFFLOAD irq_offload.c)
@@ -27,7 +28,5 @@ zephyr_library_sources_ifdef(CONFIG_CACHE_MANAGEMENT cache.c)
 if ((CONFIG_MP_NUM_CPUS GREATER 1) OR (CONFIG_SMP))
   zephyr_library_sources(smp.c)
 endif ()
-
-add_subdirectory_ifdef(CONFIG_ARM_MPU cortex_r/mpu)
 
 zephyr_cc_option_ifdef(CONFIG_USERSPACE -mno-outline-atomics)
