+++ arch/arm64/core/cortex_r/Kconfig
