--- arch/arm64/core/cortex_r/mpu/CMakeLists.txt
@@ -1,7 +1,0 @@
-# Copyright (c) 2021 Arm Limited (or its affiliates). All rights reserved.
-
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_library()
-
-zephyr_library_sources(arm_mpu.c)
