+++ soc/arm64/arm/fvp_aemv8r/arm_mpu_regions.c
@@ -5,7 +5,7 @@
 
 #include <sys/slist.h>
 #include <linker/linker-defs.h>
-#include <arch/arm64/cortex_r/mpu/arm_mpu.h>
+#include <arch/arm64/cortex_r/arm_mpu.h>
 
 
 #define DEVICE_REGION_START 0x80000000UL
